+++ src/main/kotlin/com/github/denofevil/aurelia/InspectionSuppressor.kt
@@ -9,8 +9,9 @@ import com.sixrr.inspectjs.validity.BadExpressionStatementJSInspection
 class InspectionSuppressor : JSInspectionSuppressor() {
     override fun isSuppressedFor(element: PsiElement, toolId: String): Boolean {
         if (toolId == InspectionProfileEntry.getShortName(BadExpressionStatementJSInspection::class.java.simpleName) &&
-            Aurelia.present(element.project) &&
-            FileContextUtil.getContextFile(element) != element.containingFile) {
+            Aurelia.isPresentFor(element.project) &&
+            FileContextUtil.getContextFile(element) != element.containingFile
+        ) {
             return true
         }
         return super.isSuppressedFor(element, toolId)