+++ CHANGELOG.md
@@ -2,18 +2,31 @@
 
 ## [Unreleased]
 
-### 1.3.0
-* Support for 2024.3.5
+### 1.4.0
+
+#### Features
+
+* Custom component recognition (by file name and @customElement/@containerless annotations)
+* Custom component bindable property recognition (by @bindable)
+* Suppress <template> and <require> element warnings
+
+#### Improved
+
+* Aurelia detection with large projects
 
 ### 1.2.1
+
 * Updated since build to support older versions
 
 ### 1.2.0
+
 #### Features
+
 * Support <let> element recognition
 * Support the `else` attribute
 * Adds support for `promise.bind` recognition
 
 #### Fixes
+
 * Remove deprecated getDependencies Call
 * Remove deprecated getBaseDir call